♻️ refactor: 코드래빗 리뷰 반영 - 파트 매핑 단일화, 레이어 의존 정리
PR #668 리뷰 8건 중 4건을 반영했다.

## 파트 매핑 중복 제거 (Major)

applicationStats.ts 와 evaluationStats.ts 가 PART_TAG_TO_KEY 와 emptyPartCounts 를
각자 들고 있었다. TRACK_PART_TAG(트랙 -> 파트)는 단일 소스로 뽑아 두고 정작 2단계
매핑(파트 -> 화면 키)은 두 벌이었다. 한쪽만 고치면 지원 현황과 평가 현황의 파트
집계가 서로 갈라진다.

applicantMapper 로 올리고 toPartKey(트랙 -> 화면 키) 하나로 합쳤다. 두 집계 파일은
이제 같은 함수를 부른다.

## model -> ui 의존 제거

evaluationStats(model)가 PartBreakdown 타입을 PartBreakdownTooltip(ui)에서 가져오고
있었다. 레포 전체에서 model 이 ui 를 참조하는 유일한 사례였고 이번 작업에서 생긴
것이다.

PartBreakdown 을 model/parts 로 옮겼다. 집계가 이 형태를 만들어 내므로 원래 model 에
있어야 할 타입이다. 툴팁 컴포넌트를 포함한 참조 4곳을 새 위치로 바꿨다.

## 테스트가 가드를 검증하지 못하던 문제

toCount 의 Number.isFinite 가드를 검증하려던 테스트가 빈 문자열을 쓰고 있었다.
Number("") 는 0 이라 가드가 없어도 통과한다. "abc"(NaN 이 되는 값)를 추가해 실제로
분기를 타게 하고, 빈 문자열 케이스도 함께 남겼다.

## 리스트 key

SchoolPartChartCard 의 key 가 chapterId 와 표시용 학교명 조합이었다. 이름은
shortenSchoolName 으로 축약한 값이라 겹칠 수 있다. 서버 응답에 이미 있는 schoolId 를
SchoolDatum 에 실어 key 로 쓴다.

## MAX_CHAPTERS 는 동작을 바꾸지 않고 근거만 남겼다

리뷰는 "10개 제한을 문서화하라"였는데, 확인해보니 임의 숫자가 아니라 레이아웃
제약이었다.

    카드 내부 폭  w-263(1052px) - px-8*2 = 988px
    막대 10개     82*10 + gap 10*9 + 20 = 930px  OK
    막대 11개     82*11 + gap 10*10 + 20 = 1022px  넘침

막대가 shrink-0 이고 카드에 overflow 처리가 없어 11개부터 카드 밖으로 나간다.
이름 그대로 랭킹 카드고 이미 count 내림차순으로 자르고 있어 상위 10개 표시가 의도된
동작으로 보인다. 동작은 그대로 두고 왜 10인지와 전수 확인용으로 쓰면 안 된다는 점을
주석에 남겼다.

## 반영하지 않은 것

- hasActiveGisu 플래그: 활성 기수가 없을 때 실패 문구가 나오는 건 맞지만, /gisu/active
  가 성공하면서 gisuId 만 없는 좁은 경로다. 훅 2개와 라우트 2개를 함께 고쳐야 해서
  후속으로 미룬다.
- dataUpdatedAt 0 가드: 리뷰도 "현재는 화면에 영향이 없다"고 적었다. 잠재 이슈라
  후속으로 둔다.
- 학교$ 규칙 테스트: 27개 학교 중 이 규칙에 걸리는 이름이 없어 죽은 코드일 가능성이
  있다. 테스트를 짜면 그게 드러나므로 규칙 유지 여부와 함께 판단한다.
